LayoutTests/imported/w3c/web-platform-tests/css/css-anchor-position/position-try-004-expected.txt


FAIL .target 1 assert_equals:
<div class="target" data-offset-x="0" data-expected-margin-left="0" data-expected-margin-right="10" data-expected-margin-top="10" data-expected-margin-bottom="0"></div>
offsetLeft expected 0 but got 190
FAIL .target 2 assert_equals:
<div class="target" data-offset-x="200" data-expected-margin-left="10" data-expected-margin-right="0" data-expected-margin-top="0" data-expected-margin-bottom="10"></div>
offsetLeft expected 200 but got 190
